requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
psycopg2-binary==2.9.6
python-telegram-bot==20.3
pytz==2023.3
requests==2.30.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==2.0.2
